mlops/bidlstm/predictor/requirements.txt
Flask
flask-cors
gunicorn
nilearn
google-cloud-storage